# server/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Sonarly API
  version: 2.0.0
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  description: >
    Native REST API of the Sonarly Go server. Documented routes
    are the native /api surface plus the /health, /healthz and /ready
    liveness probes;
    the OpenSubsonic adapter under /rest is a separate (Subsonic-compatible)
    contract and is intentionally not part of this spec.

    Error contract: every non-2xx response body is `{"error": "..."}` with a
    generic message for 5xx (driver errors never reach the client).
servers:
  - url: /
# Default auth: the signed sessionId cookie OR the X-API-Key header.
# Public operations override with `security: []`; /api/events overrides with
# session-cookie-only (EventSource cannot set headers).
security:
  - sessionId: []
  - apiKey: []
...
paths:
...
  /api/ratings:
    post:
      operationId: setRating
      tags: [interactions]
      summary: Rate a song, album, artist, or playlist (0..5 in 0.5 steps)
      description: >
        Upserts the caller's rating in the user_* junction row; song ratings
        recompute songs.average_rating. Passing null (or omitting the key)
        Passing null (or omitting the key) clears the rating.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: >
                Exactly one entity: a per-type id key (songId, albumId,
                artistId, playlistId) OR the legacy {entityType,
                entityId} form with entityType song|album|artist|playlist.
              properties:
                songId: { type: string }
                albumId: { type: string }
                artistId: { type: string }
                playlistId: { type: string }
                entityType:
                  type: string
                  enum: [song, album, artist, playlist]
                entityId: { type: string }
                entityID: { type: string }
                rating:
                  type: [number, 'null']
                  minimum: 0
                  maximum: 5
                  multipleOf: 0.5
                  description: 0..5 in 0.5 increments; null clears.
      responses:
        '200':
          description: Rating stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    sessionId:
      type: apiKey
      in: cookie
      name: sessionId
      description: Signed session cookie (HttpOnly, SameSite=Strict), set by /api/login and /api/setup. The only credential /api/events accepts.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Alternative to the session cookie on ordinary /api routes (not /api/events). SHA-256 digest looked up in api_keys.
...
  responses:
    BadRequest:
      description: Malformed input.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: No or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, enum: [true] }
    # --- enums ---------------------------------------------------------------
...